Move creation of OrcDataOutput to a factory
